+++ HISTORY.rst
@@ -22,3 +22,19 @@ History
 ------------------
 
 * Full requirements for keras imported
+
+0.2.3(2017-06-22)
+------------------
+
+* Fixed build_featurizer troubles with building new csv paths in current directory
+
+0.2.4(2017-06-22)
+------------------
+
+* Fixed more bugs in build_featurizer
+
+0.2.5(2017-06-22)
+------------------
+* 100% test coverage
+* Fixed a problem where a combined directory + csv was appending to the wrong
+  rows when there was a mismatch between the directory and the csv.